locator: refactor topology out of token_metadata
