+++ configure.py
@@ -945,6 +945,7 @@ scylla_core = (['message/messaging_service.cc',
                 'locator/ec2_snitch.cc',
                 'locator/ec2_multi_region_snitch.cc',
                 'locator/gce_snitch.cc',
+                'locator/topology.cc',
                 'service/client_state.cc',
                 'service/storage_service.cc',
                 'service/misc_services.cc',
